# contract/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ExpenseRequestV2" : {
        "properties" : {
          "amount" : {
            "format" : "int64",
            "maximum" : 99999999999999999,
            "minimum" : 1,
            "type" : "integer"
          },
          "category" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "currency" : {
            "maxLength" : 3,
            "minLength" : 0,
            "type" : "string"
          },
          "date" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "minLength" : 1,
            "type" : "string"
          },
          "exchangeRate" : {
            "minimum" : 1.0E-6,
            "type" : "number"
          },
          "expenseType" : {
            "type" : "string"
          },
          "reimbursable" : {
            "type" : "boolean"
          }
        },
        "required" : [ "amount", "description" ],
        "type" : "object"
      },
...
      "ExpenseResponseV2" : {
        "properties" : {
          "amount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "amountInBaseCurrency" : {
            "format" : "int64",
            "type" : "integer"
          },
          "category" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "date" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "exchangeRate" : {
            "type" : "number"
          },
          "expenseType" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "reimbursable" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Backend-issued JWT from POST /auth/login, sent as `Authorization: Bearer <token>`.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v2/expenses" : {
      "delete" : {
        "operationId" : "v2DeleteAllExpenses",
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      },
      "get" : {
        "operationId" : "v2ListExpenses",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ExpenseResponseV2"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      },
      "post" : {
        "operationId" : "v2CreateExpense",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExpenseRequestV2"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpenseResponseV2"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      }
    },
    "/api/v2/expenses/export" : {
      "get" : {
        "operationId" : "v2ExportExpenses",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "byte",
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      }
    },
    "/api/v2/expenses/import" : {
      "post" : {
        "operationId" : "v2ImportExpenses",
        "parameters" : [ {
          "in" : "query",
          "name" : "strict",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [ "file" ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImportResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      }
    },
    "/api/v2/expenses/import/template" : {
      "get" : {
        "operationId" : "v2ImportTemplate",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "byte",
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      }
    },
    "/api/v2/expenses/page" : {
      "get" : {
        "operationId" : "v2PageExpenses",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseExpenseResponseV2"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      }
    },
    "/api/v2/expenses/parse" : {
      "post" : {
        "operationId" : "v2ParseExpense",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ParseExpenseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ParsedExpenseResultV2"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      }
    },
    "/api/v2/expenses/report/category" : {
      "get" : {
        "operationId" : "v2CategoryReport",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CategoryReportItemV2"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      }
    },
    "/api/v2/expenses/report/daily" : {
      "get" : {
        "operationId" : "v2DailyReport",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DailyReportItemV2"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      }
    },
    "/api/v2/expenses/report/monthly" : {
      "get" : {
        "operationId" : "v2MonthlyReport",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/MonthlyReportItemV2"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      }
    },
    "/api/v2/expenses/report/monthly-comparison" : {
      "get" : {
        "operationId" : "v2MonthlyComparison",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MonthlyComparisonResponseV2"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      }
    },
    "/api/v2/expenses/report/top" : {
      "get" : {
        "operationId" : "v2TopTransactions",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 5,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ExpenseResponseV2"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      }
    },
    "/api/v2/expenses/{id}" : {
      "delete" : {
        "operationId" : "v2DeleteExpense",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      },
      "get" : {
        "operationId" : "v2GetExpense",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpenseResponseV2"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      },
      "put" : {
        "operationId" : "v2UpdateExpense",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExpenseRequestV2"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpenseResponseV2"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-v-2-controller" ]
      }
    },
...
    "/budgets" : {
      "delete" : {
        "operationId" : "deleteAll_3",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-controller" ]
      },
      "get" : {
        "operationId" : "list_4",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/BudgetResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-controller" ]
      },
      "post" : {
        "operationId" : "create_5",
        "parameters" : [ {
          "in" : "query",
          "name" : "force",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BudgetRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-controller" ]
      }
    },
...
    "/categories" : {
      "delete" : {
        "operationId" : "deleteAll_2",
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-controller" ]
      },
      "get" : {
        "operationId" : "list_3",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CategoryResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-controller" ]
      },
      "post" : {
        "operationId" : "create_4",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-controller" ]
      }
    },
...
    "/expenses" : {
      "delete" : {
        "operationId" : "deleteAll_1",
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      },
      "get" : {
        "operationId" : "list_2",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ExpenseResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      },
      "post" : {
        "operationId" : "create_3",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExpenseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpenseResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      }
    },
    "/expenses/export" : {
      "get" : {
        "operationId" : "export",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "byte",
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      }
    },
    "/expenses/import" : {
      "post" : {
        "operationId" : "importCsv",
        "parameters" : [ {
          "in" : "query",
          "name" : "strict",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [ "file" ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImportResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      }
    },
    "/expenses/import/template" : {
      "get" : {
        "operationId" : "importTemplate",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "byte",
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      }
    },
    "/expenses/page" : {
      "get" : {
        "operationId" : "page",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseExpenseResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      }
    },
    "/expenses/parse" : {
      "post" : {
        "operationId" : "parse",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ParseExpenseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ParsedExpenseResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      }
    },
    "/expenses/report/category" : {
      "get" : {
        "operationId" : "categoryReport",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CategoryReportItem"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      }
    },
    "/expenses/report/daily" : {
      "get" : {
        "operationId" : "dailyReport",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DailyReportItem"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      }
    },
    "/expenses/report/monthly" : {
      "get" : {
        "operationId" : "monthlyReport",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/MonthlyReportItem"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      }
    },
    "/expenses/report/monthly-comparison" : {
      "get" : {
        "operationId" : "monthlyComparison",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MonthlyComparisonResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}